DX12: encapsulate swapchain

+++ DX12/D3D12Research/D3D12/Graphics/Core/GraphicsResource.h
@@ -10,7 +10,7 @@ class ResourceState
 public:
 	ResourceState(D3D12_RESOURCE_STATES initialState = D3D12_RESOURCE_STATE_UNKNOWN) : m_CommonState(initialState), m_AllSameState(true) {}
 
-	void Set(D3D12_RESOURCE_STATES state, int subResource)
+	void Set(D3D12_RESOURCE_STATES state, uint32_t subResource)
 	{
 		if (subResource != D3D12_RESOURCE_BARRIER_ALL_SUBRESOURCES)
 		{
@@ -32,9 +32,9 @@ public:
 		}
 	}
 
-	D3D12_RESOURCE_STATES Get(int subResource) const
+	D3D12_RESOURCE_STATES Get(uint32_t subResource) const
 	{
-		check(m_AllSameState || subResource < m_ResourceStates.size());
+		check(m_AllSameState || subResource < (uint32_t)m_ResourceStates.size());
 		return m_AllSameState ? m_CommonState : m_ResourceStates[subResource];
 	}
